requirements.txt
streamlit>=1.30.0
requests>=2.28.0
openpyxl>=3.1.0
pandas>=1.5.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.20.0
